Amlogic-ce: implement Amlogic multi-dtb handling

+++ projects/Amlogic-ce/devices/Amlogic-ng/bootloader/update.sh
@@ -19,6 +19,7 @@ fi
 
 mount -o rw,remount $BOOT_ROOT
 
+DT_ID=""
 SUBDEVICE=""
 
 for arg in $(cat /proc/cmdline); do
@@ -39,39 +40,32 @@ for arg in $(cat /proc/cmdline); do
 
       if [ -f "/proc/device-tree/coreelec-dt-id" ]; then
         DT_ID=$(cat /proc/device-tree/coreelec-dt-id)
+        [ -n "$DT_ID" ] && SUBDEVICE="Generic"
       elif [ -f "/proc/device-tree/le-dt-id" ]; then
         DT_ID=$(cat /proc/device-tree/le-dt-id)
       fi
 
+      MIGRATE_DTB=""
       if [ -n "$DT_ID" ]; then
+        # modify DT_ID, SUBDEVICE and MIGRATE_DTB by dtb.conf
+        [ -f $SYSTEM_ROOT/usr/bin/convert_dtname ] && . $SYSTEM_ROOT/usr/bin/convert_dtname $DT_ID
+
         case $DT_ID in
-          *odroid_n2*)
-            SUBDEVICE="Odroid_N2"
-            ;;
           *odroid_c4*)
             SUBDEVICE="Odroid_C4"
             ;;
           *odroid_hc4*)
             SUBDEVICE="Odroid_HC4"
             ;;
-          *lepotato)
-            SUBDEVICE="LePotato"
-            ;;
           *lafrite)
             SUBDEVICE="LaFrite"
-            ;;
-          *)
-            SUBDEVICE="Generic"
             ;;
         esac
       fi
 
-      if [ -n "$DT_ID" -a -f "$SYSTEM_ROOT/usr/share/bootloader/device_trees/$DT_ID.dtb" ]; then
-        UPDATE_DTB_SOURCE="$SYSTEM_ROOT/usr/share/bootloader/device_trees/$DT_ID.dtb"
-      fi
-
-      if [ -f "$UPDATE_DTB_SOURCE" ]; then
-        echo "Updating device tree from $UPDATE_DTB_SOURCE..."
+      UPDATE_DTB_SOURCE="$SYSTEM_ROOT/usr/share/bootloader/device_trees/$DT_ID.dtb"
+      if [ -n "$DT_ID" -a -f "$UPDATE_DTB_SOURCE" ]; then
+        echo "Updating device tree with $DT_ID.dtb..."
         case $BOOT_PART in
           /dev/coreelec)
             dd if=/dev/zero of=/dev/dtb bs=256k count=1 status=none
@@ -82,6 +76,7 @@ for arg in $(cat /proc/cmdline); do
             cp -f "$UPDATE_DTB_SOURCE" "$BOOT_ROOT/dtb.img"
             ;;
         esac
+        [ -n "$MIGRATE_DTB" ] && eval $MIGRATE_DTB
       fi
 
       for all_dtb in /flash/*.dtb ; do
@@ -119,7 +114,7 @@ if [ -d $BOOT_ROOT/device_trees ]; then
 fi
 
 if [ -f $SYSTEM_ROOT/usr/share/bootloader/${SUBDEVICE}_boot.ini ]; then
-  echo "Updating boot.ini..."
+  echo "Updating boot.ini with ${SUBDEVICE}_boot.ini..."
   cp -p $SYSTEM_ROOT/usr/share/bootloader/${SUBDEVICE}_boot.ini $BOOT_ROOT/boot.ini
   sed -e "s/@BOOT_UUID@/$BOOT_UUID/" \
       -e "s/@DISK_UUID@/$DISK_UUID/" \
